changed

+++ src/main.rs
@@ -15,7 +15,7 @@ fn main() {
     let mut x = vec![0;repeat];
     if rank == root_rank {
         let v = (0..size*repeat).collect::<Vec<_>>();
-        root_process.scatter_into_root(&v[rank*repeat..(rank+1)*repeat], &mut x[..]);
+        root_process.scatter_into_root(&v[(rank*repeat) as usize..((rank+1)*repeat) as usize], &mut x[..]);
     } else {
         root_process.scatter_into(&mut x[..]);
     }
